openai-mcp-server/pyproject.toml
# openai-mcp-server/pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openai-mcp-server"
version = "0.1.3"
description = "An MCP server for interacting with OpenAI APIs (GPT, DALL-E, etc.)"
authors = [
    { name = "ParisNeo", email = "parisneoai@gmail.com" },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=0.6.0",
    "openai>=1.0.0",
    "python-dotenv>=0.20.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.95.0",
    "httpx>=0.24.0",
    "ascii-colors>=0.5.5",
]

[project.scripts]
openai-mcp-server = "openai_mcp_server.server:main_cli"

[project.urls]
Homepage = "https://github.com/PArisNeoMCPServers/openai-mcp-server"
Repository = "https://github.com/PArisNeoMCPServers/openai-mcp-server"